WIP removing MainActivity by moving responsibilities elsewhere/
